+++ examples/sse/script.js
@@ -1,0 +1,5 @@
+var eventSource = new EventSource('sse');
+
+eventSource.onmessage = function(event) {
+    console.log('Message from server ', event.data);
+}